Carousel + HTML +CSS
Added footers to each page, worked on contact.css some but the problem is still occurring. Worked on the index file as well and added some small media query styles for the images in the about section. Also created and worked on a JS file using ES6 for the carousel, but it's not working properly.